+++ src/rgraph_ba.cpp
@@ -6,7 +6,7 @@ using namespace Rcpp;
 // [[Rcpp::export]]
 arma::sp_mat rgraph_ba_cpp(
     const arma::sp_mat & graph,
-    const arma::colvec & dgr, int m = 1, int t = 10) {
+    const arma::colvec & dgr, int m = 1, int t = 10, bool self=true) {
 
   // Creating the empty graph
   // m0: Size of the graph (changes throughout time).
@@ -23,23 +23,27 @@ arma::sp_mat rgraph_ba_cpp(
 
   // Start the process, K is sum(dgr)
   int K = sum(dgr);
-  // std::cout << dgr ;
 
-  int m1;
+  int m0_trunc;
   double randdraw, cump;
+
+  // If self=true, then the prob are computed over m0+1, otherwise only over m0
+  int extra = self? 1 : 0;
   for(int i=0;i<t;i++) {
     // Checling user interrup
     if (i % 1000 == 0)
       Rcpp::checkUserInterrupt();
 
     // The number of conections is trucated by the number of vertices in the graph
-    // so -m1- is actually -m-, but if currently there are less vertices than m,
+    // so -m0_trunc- is actually -m-, but if currently there are less vertices than m,
     // then its truncated.
-    m1 = (m > m0)? m0 : m;
-    // if (m > m0) m1 = m0;
+    m0_trunc = (m > m0)? m0 : m;
+    // if (m > m0) m0_trunc = m0;
 
-    for (int j=0;j<m1;j++) {
-      // Incrementing the degree of the first
+    int K0 = K;
+    for (int j=0;j<m0_trunc;j++) {
+      // Incrementing the degree of the one that is been added
+      // one by one until having degree m0_trunk.
       dgr_new.at(m0) += 1.0;
 
       // std::cout << j << " Iter, "<< m << " m\n";
@@ -49,9 +53,13 @@ arma::sp_mat rgraph_ba_cpp(
       // Calculating probabilities of been drawn. -cump- is the cumsum of the
       // probabilities
       cump = 0.0;
-      for (int k=0; k<(m0+1); k++) {
+      for (int k=0; k<m0+extra; k++) {
 
-        cump += dgr_new.at(k)/(K + 1);
+        // In the case that in iter i the total degree is zero (no links)
+        // then all individuals are equally likely to receive a new link.
+        if (K0 != 0) cump += dgr_new.at(k)/(K + extra);
+        else cump += 1/m0;
+
         // if (cump > 1) printf("over1 (%04d/%04d): %9.4g\n", k+1, m0+1,cump );
 
         // Links to the set of previous vertices
@@ -72,14 +80,22 @@ arma::sp_mat rgraph_ba_cpp(
 }
 
 // [[Rcpp::export]]
-arma::sp_mat rgraph_ba_new_cpp(int m0 = 1, int m = 1, int t = 10) {
+arma::sp_mat rgraph_ba_new_cpp(int m0 = 1, int m = 1, int t = 10, bool self=true) {
   int n = m0;
-  arma::sp_mat graph(n, n);
-  graph.diag() = arma::ones(n);
 
-  arma::colvec dgr(n, arma::fill::ones);
-  dgr = dgr*2.0;
-  return rgraph_ba_cpp(graph, dgr, m, t);
+  arma::sp_mat graph(n, n);
+  arma::colvec dgr(n);
+
+  if (!self) {
+    graph.diag().fill(0.0);
+    dgr.fill(0.0);
+  }
+  else {
+    graph.diag().fill(1.0);
+    dgr.fill(2.0);
+  }
+
+  return rgraph_ba_cpp(graph, dgr, m, t, self);
 }
 
 // [[Rcpp::export]]