+++ src/rgraph_ba.h
@@ -9,7 +9,7 @@ using namespace Rcpp;
 
 arma::sp_mat rgraph_ba_cpp(
     arma::sp_mat graph,
-    arma::colvec dgr, int m = 1, int t = 10);
+    arma::colvec dgr, int m = 1, int t = 10, bool self=true);
 
-arma::sp_mat rgraph_ba_new_cpp(int m0 = 1, int m = 1, int t = 10);
+arma::sp_mat rgraph_ba_new_cpp(int m0 = 1, int m = 1, int t = 10, bool self=true);
 #endif