+++ tests/testthat/test-struct_test.R
@@ -45,3 +45,22 @@ test_that("struct_test should be reproducible (parallel version)", {
   expect_equal(b1$mean_t, b2$mean_t, tolerance=1e-2)
 
 })
+
+
+test_that("Methods of struct test", {
+  set.seed(1122)
+  x <- rdiffnet(100, 4, "central")
+  diffnet.toa(x) <- sample(x$toa, nnodes(x))
+
+  ans1 <- struct_test(x, function(g) mean(threshold(g), na.rm = TRUE), 100)
+  ans2 <- struct_test(x, function(g) mean(threshold(g), na.rm = TRUE), 100)
+
+  # pvalues of concatenated test should be the same no matter the order
+  expect_equal(c(ans1,ans2)$p.value,c(ans2,ans1)$p.value)
+
+  # The number of repetitions should increase
+  expect_output(print(c(ans2,ans1,ans1)), "300")
+
+  # Just printing
+  hist(ans1)
+})